fix: Avoid using keyword arguments in sidekiq middleware
Sidekiqs middleware chain does pass keyword arguments correctly to the
middleware objects when initializing the chain. This results in hard
exceptions in Ruby 3.0.

This CL changes the telegraf middleware for sidekiq to not use pure
keyword arguments but to accept an agent instance and an options hash as
positional arguments.

Closes #14